requirements.txt
fastapi==0.95.0
pydantic==1.10.5
mlflow==2.5.0
pandas==1.5.3
scikit-learn==1.2.2
xgboost==1.7.5
pytest==7.2.2
scipy==1.9.3
python-dotenv==0.21.0
pyyaml==6.0
matplotlib==3.7.1
seaborn==0.11.2
shap==0.41.0
uvicorn==0.34.2



